Update Ch3_IfElifElse
In Mike Pirnat's "How to Make Mistakes in Python": (http://www.oreilly.com/programming/free/files/how-to-make-mistakes-in-python.pdf) he suggests using the Strategy pattern in order to avoid large unruly If/else blocks. I remembered that I had a file where this occurred and wanted to create a simple example of this pattern to refer to in the future. More research is required